requirements.txt
streamlit==1.20.0
streamlit-aggrid==0.3.4.post3
streamlit-webrtc==0.45.0
urllib3==1.26.14
Pillow==9.1.1
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.18.0
lxml==4.9.0
numpy==1.23.5
pandas==1.5.3
bs4==0.0.1
beautifulsoup4==4.11.1
wordcloud==1.8.1
jieba==0.42.1
matplotlib==3.5.2
snownlp==0.12.3
matplotlib-inline==0.1.3